+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Changelog for the docker image are [here](/docker_CHANGELOG.md).
 ### Fixed
 
 - avoid context cache collisions by including query shape options in cache keys (#1934)
+- cap DataLoader caches and discarded loader list (#1935)
 
 ## [0.2.6]